+++ tool/travis.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
 
 # Fast fail the script on failures.
-set -e
+# set -e
 
 # Install dart_coveralls; gather and send coverage data.
-if [ "$COVERALLS_TOKEN" ] && [ "$TRAVIS_DART_VERSION" = "dev" ]; then
+if [ "$COVERALLS_TOKEN" ] && [ "$TRAVIS_DART_VERSION" = "stable" ]; then
   pub global activate dart_coveralls
-  # git clone https://github.com/bkonyi/dart-coveralls.git
-  # cd dart-coveralls
-  # pub get
-  # cd ..
+  #git clone https://github.com/bkonyi/dart-coveralls.git
+  #cd dart-coveralls
+  #pub get
+  #cd ..
   echo "Running coverage..."
   #dart dart-coveralls/bin/dart_coveralls.dart report \
   dart_coveralls report \